@@ -1,6 +1,6 @@
 {
   "name": "telepathy-agent",
-  "version": "1.1.1",
+  "version": "1.2.0",
   "author": "",
   "license": "MIT",
   "scripts": {
@@ -20,12 +20,15 @@
     "binary:linux:arm64": "npm run build && pkg dist/app.js --target node12-linux-arm64 --output ./dist/telepathy-agent-linux-arm64"
   },
   "dependencies": {
-    "@opentelemetry/api": "^1.1.0",
-    "@opentelemetry/exporter-trace-otlp-http": "^0.30.0",
-    "@opentelemetry/instrumentation-http": "^0.30.0",
-    "@opentelemetry/resources": "^1.4.0",
-    "@opentelemetry/sdk-node": "^0.30.0",
-    "@opentelemetry/semantic-conventions": "^1.4.0",
+    "@opentelemetry/api": "^1.2.0",
+    "@opentelemetry/auto-instrumentations-node": "^0.32.1",
+    "@opentelemetry/exporter-trace-otlp-http": "^0.33.0",
+    "@opentelemetry/id-generator-aws-xray": "^1.1.0",
+    "@opentelemetry/resources": "^1.7.0",
+    "@opentelemetry/sdk-node": "^0.33.0",
+    "@opentelemetry/sdk-trace-base": "^1.7.0",
+    "@opentelemetry/sdk-trace-web": "^1.7.0",
+    "@opentelemetry/semantic-conventions": "^1.7.0",
     "axios": "^0.21.1",
     "fs-extra": "^7.0.1",
     "lodash": "^4.17.20",
@@ -34,18 +37,18 @@
   },
   "devDependencies": {
     "@types/fs-extra": "^9.0.13",
-    "@types/jest": "^28.1.6",
-    "@types/lodash": "^4.14.182",
-    "@types/node": "^18.0.6",
-    "@types/sinon": "^10.0.12",
-    "@typescript-eslint/eslint-plugin": "^5.30.5",
-    "@typescript-eslint/parser": "^5.30.6",
-    "eslint": "^8.20.0",
-    "jest": "^28.1.3",
+    "@types/jest": "^29.0.3",
+    "@types/lodash": "^4.14.185",
+    "@types/node": "^18.7.18",
+    "@types/sinon": "^10.0.13",
+    "@typescript-eslint/eslint-plugin": "^5.38.0",
+    "@typescript-eslint/parser": "^5.38.0",
+    "eslint": "^8.23.1",
+    "jest": "^29.0.3",
     "pkg": "^5.8.0",
     "sinon": "^14.0.0",
-    "ts-jest": "^28.0.7",
+    "ts-jest": "^29.0.1",
     "ts-node": "^10.9.1",
-    "typescript": "^4.7.4"
+    "typescript": "^4.8.3"
   }
 }
