@@ -1,6 +1,6 @@
 {
   "name": "telepathy-server",
-  "version": "1.3.0",
+  "version": "1.4.0",
   "author": "",
   "license": "MIT",
   "scripts": {
@@ -11,7 +11,7 @@
     "test": "jest --coverage || true",
     "test:dev": "jest --watch --coverage",
     "////": "-----------------------------------------------------------------",
-    "dev": "ts-node --require './src/utils-std-ts/tracing.ts' ./src/app.ts",
+    "dev": "ts-node ./src/app.ts",
     "dev:docker": "docker-packaging-config/start-dev-docker.sh || true",
     "dev:test": "NODE_ENV=dev jest --watch --config='jest.config.js'",
     "binary:alpine:x64": "npm run build && pkg dist/app.js --target node16-alpine-x64 --output ./dist/telepathy-server-linux-alpine",
@@ -21,12 +21,15 @@
   },
   "dependencies": {
     "@fastify/cors": "^8.0.0",
-    "@opentelemetry/api": "^1.1.0",
-    "@opentelemetry/exporter-trace-otlp-http": "^0.30.0",
-    "@opentelemetry/instrumentation-http": "^0.30.0",
-    "@opentelemetry/resources": "^1.4.0",
-    "@opentelemetry/sdk-node": "^0.30.0",
-    "@opentelemetry/semantic-conventions": "^1.4.0",
+    "@opentelemetry/api": "^1.2.0",
+    "@opentelemetry/auto-instrumentations-node": "^0.32.1",
+    "@opentelemetry/exporter-trace-otlp-http": "^0.33.0",
+    "@opentelemetry/id-generator-aws-xray": "^1.1.0",
+    "@opentelemetry/resources": "^1.7.0",
+    "@opentelemetry/sdk-node": "^0.33.0",
+    "@opentelemetry/sdk-trace-base": "^1.7.0",
+    "@opentelemetry/sdk-trace-web": "^1.7.0",
+    "@opentelemetry/semantic-conventions": "^1.7.0",
     "bcrypt": "^5.0.1",
     "body-parser": "^1.19.0",
     "fastify": "^4.2.0",
@@ -40,16 +43,16 @@
   },
   "devDependencies": {
     "@types/fs-extra": "^9.0.13",
-    "@types/jest": "^28.1.6",
-    "@types/lodash": "^4.14.182",
-    "@types/node": "^18.0.6",
-    "@typescript-eslint/eslint-plugin": "^5.30.5",
-    "@typescript-eslint/parser": "^5.30.6",
-    "eslint": "^8.20.0",
-    "jest": "^28.1.3",
+    "@types/jest": "^29.0.3",
+    "@types/lodash": "^4.14.185",
+    "@types/node": "^18.7.18",
+    "@typescript-eslint/eslint-plugin": "^5.38.0",
+    "@typescript-eslint/parser": "^5.38.0",
+    "eslint": "^8.23.1",
+    "jest": "^29.0.3",
     "sinon": "^14.0.0",
-    "ts-jest": "^28.0.7",
+    "ts-jest": "^29.0.1",
     "ts-node": "^10.9.1",
-    "typescript": "^4.7.4"
+    "typescript": "^4.8.3"
   }
 }
