@@ -1,8 +1,8 @@
 {
   "name": "telepathy-web",
-  "version": "1.1.1",
+  "version": "1.1.2",
   "scripts": {
-    "dev": "vite",
+    "dev": "vite --port 3000",
     "build": "vue-tsc --noEmit && vite build",
     "preview": "vite preview --port 5050",
     "test": "vitest --environment jsdom || true",
@@ -17,29 +17,29 @@
     "jwt-decode": "^3.1.2",
     "lodash": "^4.17.21",
     "mitt": "^3.0.0",
-    "pinia": "^2.0.14",
+    "pinia": "^2.0.17",
     "vue": "^3.2.37",
-    "vue-router": "^4.1.1"
+    "vue-router": "^4.1.3"
   },
   "devDependencies": {
     "@rushstack/eslint-patch": "^1.1.4",
-    "@types/jsdom": "^16.2.14",
-    "@types/node": "^18.0.3",
-    "@vitejs/plugin-vue": "^2.3.3",
+    "@types/jsdom": "^20.0.0",
+    "@types/node": "^18.6.2",
+    "@vitejs/plugin-vue": "^3.0.1",
     "@vue/eslint-config-prettier": "^7.0.0",
     "@vue/eslint-config-typescript": "^11.0.0",
     "@vue/test-utils": "^2.0.2",
     "@vue/tsconfig": "^0.1.3",
-    "cypress": "^10.3.0",
-    "eslint": "^8.19.0",
+    "cypress": "^10.3.1",
+    "eslint": "^8.20.0",
     "eslint-plugin-cypress": "^2.12.1",
-    "eslint-plugin-vue": "^9.2.0",
+    "eslint-plugin-vue": "^9.3.0",
     "jsdom": "^20.0.0",
     "prettier": "^2.7.1",
     "start-server-and-test": "^1.14.0",
     "typescript": "~4.7.4",
-    "vite": "^2.9.14",
-    "vitest": "^0.17.1",
-    "vue-tsc": "^0.38.3"
+    "vite": "^3.0.4",
+    "vitest": "^0.19.1",
+    "vue-tsc": "^0.39.2"
   }
 }
