@@ -1,20 +1,15 @@
-import * as path from "path";
 import * as cron from "node-cron";
-import { Logger } from "../utils-std-ts/logger";
 import { FastifyInstance, RequestGenericInterface } from "fastify";
 import { AppContext } from "../appContext";
 import { Auth } from "../data/auth";
 import { Task } from "../common-model/task";
-const opentelemetry = require("@opentelemetry/api");
-
-const logger = new Logger(path.basename(__filename));
+import { StandardTracer } from "../utils-std-ts/standardTracer";
 
 async function routes(fastify: FastifyInstance): Promise<void> {
   //
   fastify.get("/", async (req, res) => {
-    logger.debug(`[${req.method}] ${req.url}`);
     await Auth.mustBeAuthenticated(req, res);
-    const tasks = await AppContext.getTasks().list();
+    const tasks = await AppContext.getTasks().list(StandardTracer.getSpanFromRequest(req));
     res.status(200).send({
       tasks,
     });
@@ -28,7 +23,6 @@ async function routes(fastify: FastifyInstance): Promise<void> {
     };
   }
   fastify.post<Post>("/", async (req, res) => {
-    logger.info(`[${req.method}] ${req.url}`);
     Auth.mustBeAuthenticated(req, res);
     if (!req.body.name) {
       return res.status(400).send({ error: "Missing: Name" });
@@ -41,8 +35,8 @@ async function routes(fastify: FastifyInstance): Promise<void> {
     }
 
     const newTask = Task.fromJson(req.body);
-    await AppContext.getTasks().add(newTask);
-    AppContext.getScheduler().calculate();
+    await AppContext.getTasks().add(StandardTracer.getSpanFromRequest(req), newTask);
+    AppContext.getScheduler().calculate(StandardTracer.getSpanFromRequest(req));
     res.status(201).send(newTask.toJson());
   });
 }
