@@ -1,18 +1,27 @@
 import { watchFile } from "fs-extra";
 import { Auth } from "./process/auth";
-import { config } from "./config";
+import { Config } from "./config";
 import { TaskExecutions } from "./process/taskExecutions";
 import { Logger } from "./utils-std-ts/logger";
+import { StandardTracer } from "./utils-std-ts/standardTracer";
+import { AppContext } from "./appContext";
 
 const logger = new Logger("app");
 
 logger.info(`====== Starting Telepathy Agent ======`);
 
 Promise.resolve().then(async () => {
-  watchFile(config.CONFIG_FILE, () => {
-    logger.info(`Config updated: ${config.CONFIG_FILE}`);
-    config.reload();
+  //
+  const config = new Config();
+  await config.reload();
+  AppContext.setConfig(config);
+  watchFile(AppContext.getConfig().CONFIG_FILE, () => {
+    logger.info(`Config updated: ${AppContext.getConfig().CONFIG_FILE}`);
+    AppContext.getConfig().reload();
   });
+
+  StandardTracer.initTelemetry();
+
   setTimeout(() => {
     Auth.check();
   }, 100);
