@@ -1,5 +1,5 @@
 import { Auth } from "./auth";
 
 test("get token if not empty", async () => {
-  const header = await Auth.getAuthHeader();
+  const header = await Auth.getAuthHeader(null);
 });
