@@ -1,10 +1,13 @@
+import { Span } from "@opentelemetry/sdk-trace-base";
 import * as fs from "fs-extra";
 import * as _ from "lodash";
 import { AppContext } from "../appContext";
 import { TaskExecution } from "../common-model/taskExecution";
 import { TaskExecutionStatus } from "../common-model/taskExecutionStatus";
 import { TaskOutput } from "../common-model/taskOutput";
+import { StandardTracer } from "../utils-std-ts/standardTracer";
 import { Logger } from "../utils-std-ts/logger";
+import { FileDBUtils } from "./fileDbUtils";
 
 const logger = new Logger("data/taskExecution");
 
@@ -12,18 +15,11 @@ export class TaskExecutions {
   //
   public taskExecutions: TaskExecution[];
 
-  public async load(): Promise<void> {
-    if (fs.existsSync(`${AppContext.getConfig().DATA_DIR}/task-executions.json`)) {
-      await fs.readJSON(`${AppContext.getConfig().DATA_DIR}/task-executions.json`).then((data) => {
-        this.taskExecutions = data;
-      });
-    } else {
-      this.taskExecutions = [];
-      await this.save();
-    }
+  public async load(context: Span): Promise<void> {
+    this.taskExecutions = await FileDBUtils.load(context, "task-executions", []);
   }
 
-  public async get(id: string): Promise<TaskExecution> {
+  public async get(context: Span, id: string): Promise<TaskExecution> {
     return TaskExecution.fromJson(
       _.find(this.taskExecutions, {
         id,
@@ -31,22 +27,25 @@ export class TaskExecutions {
     );
   }
 
-  public async delete(id: string): Promise<void> {
+  public async delete(context: Span, id: string): Promise<void> {
+    const span = StandardTracer.startSpan("TaskExecutions_delete", context);
     const position = _.findIndex(this.taskExecutions, {
       id,
     });
     if (position >= 0) {
-      await this.deleteLogs(this.taskExecutions[position].taskId, id);
+      await this.deleteLogs(span, this.taskExecutions[position].taskId, id);
       this.taskExecutions.splice(position, 1);
     }
-    await this.save();
+    await this.save(span);
+    span.end();
   }
 
-  public async list(): Promise<TaskExecution[]> {
+  public async list(context: Span): Promise<TaskExecution[]> {
     return this.taskExecutions;
   }
 
-  public async update(id: string, taskExecutionUpdate: TaskExecution): Promise<void> {
+  public async update(context: Span, id: string, taskExecutionUpdate: TaskExecution): Promise<void> {
+    const span = StandardTracer.startSpan("TaskExecutions_update", context);
     const taskExecution = _.find(this.taskExecutions, {
       id,
     });
@@ -58,13 +57,13 @@ export class TaskExecutions {
     taskExecution.dateExecuted = taskExecutionUpdate.dateExecuted;
     taskExecution.dateAgentAlive = taskExecutionUpdate.dateAgentAlive;
     taskExecution.outputs = taskExecutionUpdate.outputs;
-    await this.save();
+    await this.save(span);
     if (taskExecution.status === TaskExecutionStatus.executed) {
       // Delay to wait for the log to be stable
       setTimeout(async () => {
         taskExecution.outputs = [];
-        const task = await AppContext.getTasks().get(taskExecution.taskId);
-        const logs = await this.getLogs(taskExecution.id, taskExecution.taskId);
+        const task = await AppContext.getTasks().get(span, taskExecution.taskId);
+        const logs = await this.getLogs(span, taskExecution.id, taskExecution.taskId);
         for (const outputDefinition of task.outputDefinitions) {
           try {
             const regex = new RegExp(outputDefinition.pattern);
@@ -76,8 +75,9 @@ export class TaskExecutions {
               taskOutput.taskOutputDefinitionId = outputDefinition.id;
               taskExecution.outputs.push(taskOutput);
             }
-            await this.save();
+            await this.save(span);
           } catch (error) {
+            span.recordException(error);
             logger.error(`Output Pattern Matching Failed: ${error}`);
           }
         }
@@ -85,8 +85,9 @@ export class TaskExecutions {
     }
   }
 
-  public async createFromTaskId(taskId: string): Promise<TaskExecution> {
-    const task = await AppContext.getTasks().get(taskId);
+  public async createFromTaskId(context: Span, taskId: string): Promise<TaskExecution> {
+    const span = StandardTracer.startSpan("TaskExecutions_createFromTaskId", context);
+    const task = await AppContext.getTasks().get(span, taskId);
     const newTaskExecution = new TaskExecution();
     newTaskExecution.taskId = taskId;
     newTaskExecution.script = task.script;
@@ -94,11 +95,12 @@ export class TaskExecutions {
     newTaskExecution.status = TaskExecutionStatus.queued;
     newTaskExecution.dateQueued = new Date();
     this.taskExecutions.push(newTaskExecution);
-    await this.save();
+    await this.save(span);
+    span.end();
     return newTaskExecution;
   }
 
-  public async getLogs(id: string, taskId: string): Promise<Buffer> {
+  public async getLogs(context: Span, id: string, taskId: string): Promise<Buffer> {
     if (fs.existsSync(`${AppContext.getConfig().DATA_DIR}/logs/${taskId}_${id}.log`)) {
       return await fs.readFile(`${AppContext.getConfig().DATA_DIR}/logs/${taskId}_${id}.log`);
     } else {
@@ -106,18 +108,20 @@ export class TaskExecutions {
     }
   }
 
-  public async deleteLogs(id: string, taskId: string): Promise<void> {
+  public async deleteLogs(context: Span, id: string, taskId: string): Promise<void> {
     if (fs.existsSync(`${AppContext.getConfig().DATA_DIR}/logs/${taskId}_${id}.log`)) {
       await fs.remove(`${AppContext.getConfig().DATA_DIR}/logs/${taskId}_${id}.log`);
     }
   }
 
-  public async updateLogs(taskExecutionId: string, taskId: string, logs: Buffer): Promise<void> {
+  public async updateLogs(context: Span, taskExecutionId: string, taskId: string, logs: Buffer): Promise<void> {
+    const span = StandardTracer.startSpan("TaskExecutions_updateLogs", context);
     await fs.ensureDir(`${AppContext.getConfig().DATA_DIR}/logs`);
     await fs.writeFile(`${AppContext.getConfig().DATA_DIR}/logs/${taskId}_${taskExecutionId}.log`, logs);
+    span.end();
   }
 
-  public async save(): Promise<void> {
-    await fs.writeJSON(`${AppContext.getConfig().DATA_DIR}/task-executions.json`, this.taskExecutions);
+  public async save(context: Span): Promise<void> {
+    await FileDBUtils.save(context, "task-executions", this.taskExecutions);
   }
 }
