@@ -1,24 +1,18 @@
-import * as fs from "fs-extra";
+import { Span } from "@opentelemetry/sdk-trace-base";
 import * as _ from "lodash";
-import { AppContext } from "../appContext";
 import { Task } from "../common-model/task";
+import { StandardTracer } from "../utils-std-ts/standardTracer";
+import { FileDBUtils } from "./fileDbUtils";
 
 export class Tasks {
   //
   public tasks: Task[];
 
-  public async load(): Promise<void> {
-    if (fs.existsSync(`${AppContext.getConfig().DATA_DIR}/tasks.json`)) {
-      fs.readJSON(`${AppContext.getConfig().DATA_DIR}/tasks.json`).then((data) => {
-        this.tasks = data;
-      });
-    } else {
-      this.tasks = [];
-      await this.save();
-    }
+  public async load(context: Span): Promise<void> {
+    this.tasks = await FileDBUtils.load(context, "tasks", []);
   }
 
-  public async get(id: string): Promise<Task> {
+  public async get(context: Span, id: string): Promise<Task> {
     return Task.fromJson(
       _.find(this.tasks, {
         id,
@@ -26,7 +20,8 @@ export class Tasks {
     );
   }
 
-  public async update(id: string, taskUpdate: Task): Promise<void> {
+  public async update(context: Span, id: string, taskUpdate: Task): Promise<void> {
+    const span = StandardTracer.startSpan("Tasks_update", context);
     const task = _.find(this.tasks, {
       id,
     }) as Task;
@@ -36,29 +31,32 @@ export class Tasks {
     task.tag = taskUpdate.tag;
     task.webhook = taskUpdate.webhook;
     task.outputDefinitions = taskUpdate.outputDefinitions;
-    await this.save();
+    await this.save(span);
+    span.end();
   }
 
-  public async delete(id: string): Promise<void> {
+  public async delete(context: Span, id: string): Promise<void> {
+    const span = StandardTracer.startSpan("Tasks_delete", context);
     const position = _.findIndex(this.tasks, {
       id,
     });
     if (position >= 0) {
       this.tasks.splice(position, 1);
     }
-    await this.save();
+    await this.save(span);
+    span.end();
   }
 
-  public async list(): Promise<Task[]> {
+  public async list(context: Span): Promise<Task[]> {
     return _.cloneDeep(this.tasks);
   }
 
-  public async add(task: Task): Promise<void> {
+  public async add(context: Span, task: Task): Promise<void> {
     this.tasks.push(task);
-    await this.save();
+    await this.save(context);
   }
 
-  public async save(): Promise<void> {
-    await fs.writeJSON(`${AppContext.getConfig().DATA_DIR}/tasks.json`, this.tasks);
+  public async save(context: Span): Promise<void> {
+    await FileDBUtils.save(context, "tasks", this.tasks);
   }
 }
