@@ -4,14 +4,14 @@ import { UserPassword } from "./userPassword";
 test("Password should be successfully verified if it's the same", async () => {
   const password = "testPassword1234";
   const user = new User();
-  await UserPassword.setPassword(user, password);
-  expect(await UserPassword.checkPassword(user, password)).toBeTruthy();
+  await UserPassword.setPassword(null, user, password);
+  expect(await UserPassword.checkPassword(null, user, password)).toBeTruthy();
 });
 
 test("Password should be faile to be verified if it's not the same", async () => {
   const password = "testPassword1234";
   const passwordWrong = "testPassword12345";
   const user = new User();
-  await UserPassword.setPassword(user, password);
-  expect(await UserPassword.checkPassword(user, passwordWrong)).toBeFalsy();
+  await UserPassword.setPassword(null, user, password);
+  expect(await UserPassword.checkPassword(null, user, passwordWrong)).toBeFalsy();
 });
