@@ -1,14 +1,11 @@
 import * as _ from "lodash";
-import * as path from "path";
-import { Logger } from "../utils-std-ts/logger";
 import { FastifyInstance, RequestGenericInterface } from "fastify";
 import { AppContext } from "../appContext";
 import { Auth } from "../data/auth";
 import { Agent } from "../common-model/agent";
 import { User } from "../common-model/user";
 import { TaskExecutionStatus } from "../common-model/taskExecutionStatus";
-
-const logger = new Logger(path.basename(__filename));
+import { StandardTracer } from "../utils-std-ts/standardTracer";
 
 async function routes(fastify: FastifyInstance): Promise<void> {
   //
@@ -22,7 +19,6 @@ async function routes(fastify: FastifyInstance): Promise<void> {
     };
   }
   fastify.post<PostSession>("/session", async (req, res) => {
-    logger.debug(`[${req.method}] ${req.url}`);
     if (!req.body.key) {
       return res.status(400).send({ error: "Missing: Key" });
     } else if (req.body.key !== AppContext.getConfig().AGENT_KEY) {
@@ -32,7 +28,7 @@ async function routes(fastify: FastifyInstance): Promise<void> {
       if (req.body.tags) {
         agent.tags = req.body.tags;
       }
-      await AppContext.getAgents().register(agent);
+      await AppContext.getAgents().register(StandardTracer.getSpanFromRequest(req), agent);
       const userAgent = new User();
       userAgent.name = req.params.agentId;
       res.status(201).send({ success: true, token: await Auth.generateJWT(userAgent) });
@@ -45,12 +41,14 @@ async function routes(fastify: FastifyInstance): Promise<void> {
     };
   }
   fastify.get<GetTaskExecutions>("/tasks/executions", async (req, res) => {
-    logger.debug(`[${req.method}] ${req.url}`);
     await Auth.mustBeAuthenticated(req, res);
-    const taskExecutionsQueued = _.filter(await AppContext.getTaskExecutions().list(), {
-      status: TaskExecutionStatus.queued,
-    });
-    const agent = await AppContext.getAgents().get(req.params.agentId);
+    const taskExecutionsQueued = _.filter(
+      await AppContext.getTaskExecutions().list(StandardTracer.getSpanFromRequest(req)),
+      {
+        status: TaskExecutionStatus.queued,
+      }
+    );
+    const agent = await AppContext.getAgents().get(StandardTracer.getSpanFromRequest(req), req.params.agentId);
     const taskExecutionCompatible = [];
     for (const taskExecution of taskExecutionsQueued) {
       if (!taskExecution.tag || agent.tags.indexOf(taskExecution.tag) >= 0) {
