@@ -1,11 +1,8 @@
-import * as path from "path";
-import { Logger } from "../utils-std-ts/logger";
 import { FastifyInstance, RequestGenericInterface } from "fastify";
 import { Auth } from "../data/auth";
 import { AppContext } from "../appContext";
 import { TaskExecution } from "../common-model/taskExecution";
-
-const logger = new Logger(path.basename(__filename));
+import { StandardTracer } from "../utils-std-ts/standardTracer";
 
 async function routes(fastify: FastifyInstance): Promise<void> {
   //
@@ -16,11 +13,17 @@ async function routes(fastify: FastifyInstance): Promise<void> {
     };
   }
   fastify.get<GetAgentExecutionId>("/:taskExecutionId", async (req, res) => {
-    logger.info(`[${req.method}] ${req.url}`);
     Auth.mustBeAuthenticated(req, res);
-    const taskExecution = await AppContext.getTaskExecutions().get(req.params.taskExecutionId);
+    const taskExecution = await AppContext.getTaskExecutions().get(
+      StandardTracer.getSpanFromRequest(req),
+      req.params.taskExecutionId
+    );
     taskExecution.dateAgentAlive = new Date();
-    await AppContext.getTaskExecutions().update(req.params.taskExecutionId, taskExecution);
+    await AppContext.getTaskExecutions().update(
+      StandardTracer.getSpanFromRequest(req),
+      req.params.taskExecutionId,
+      taskExecution
+    );
     res.status(200).send(taskExecution);
   });
 
@@ -31,10 +34,13 @@ async function routes(fastify: FastifyInstance): Promise<void> {
     };
   }
   fastify.put<PutAgentExecutionId>("/:taskExecutionId", async (req, res) => {
-    logger.info(`[${req.method}] ${req.url}`);
     Auth.mustBeAuthenticated(req, res);
     const taskExecutionUpdate = TaskExecution.fromJson(req.body);
-    await AppContext.getTaskExecutions().update(req.params.taskExecutionId, taskExecutionUpdate);
+    await AppContext.getTaskExecutions().update(
+      StandardTracer.getSpanFromRequest(req),
+      req.params.taskExecutionId,
+      taskExecutionUpdate
+    );
     res.status(200).send({});
   });
 
@@ -48,15 +54,26 @@ async function routes(fastify: FastifyInstance): Promise<void> {
     };
   }
   fastify.put<PutAgentExecutionIdLog>("/:taskExecutionId/logs", async (req, res) => {
-    logger.info(`[${req.method}] ${req.url}`);
     Auth.mustBeAuthenticated(req, res);
     if (!req.body.logs) {
       return res.status(400).send({ error: "Missing: Logs" });
     }
-    const taskExecution = await AppContext.getTaskExecutions().get(req.params.taskExecutionId);
+    const taskExecution = await AppContext.getTaskExecutions().get(
+      StandardTracer.getSpanFromRequest(req),
+      req.params.taskExecutionId
+    );
     taskExecution.dateAgentAlive = new Date();
-    await AppContext.getTaskExecutions().update(req.params.taskExecutionId, taskExecution);
-    await AppContext.getTaskExecutions().updateLogs(req.params.taskExecutionId, req.params.taskId, req.body.logs);
+    await AppContext.getTaskExecutions().update(
+      StandardTracer.getSpanFromRequest(req),
+      req.params.taskExecutionId,
+      taskExecution
+    );
+    await AppContext.getTaskExecutions().updateLogs(
+      StandardTracer.getSpanFromRequest(req),
+      req.params.taskExecutionId,
+      req.params.taskId,
+      req.body.logs
+    );
     res.status(200).send({});
   });
 }
