@@ -1,10 +1,7 @@
 import * as _ from "lodash";
-import * as path from "path";
-import { Logger } from "../utils-std-ts/logger";
 import { FastifyInstance, RequestGenericInterface } from "fastify";
 import { AppContext } from "../appContext";
-
-const logger = new Logger(path.basename(__filename));
+import { StandardTracer } from "../utils-std-ts/standardTracer";
 
 async function routes(fastify: FastifyInstance): Promise<void> {
   //
@@ -14,15 +11,17 @@ async function routes(fastify: FastifyInstance): Promise<void> {
     };
   }
   fastify.post<PostTaskWebhook>("/:webhookId", async (req, res) => {
-    logger.info(`[${req.method}] ${req.url}`);
-    const tasks = await AppContext.getTasks().list();
+    const tasks = await AppContext.getTasks().list(StandardTracer.getSpanFromRequest(req));
     const task = _.find(tasks, {
       webhook: req.params.webhookId,
     });
     if (!task) {
       return res.status(404).send({ error: "Not Found" });
     }
-    const newTaskExecution = await AppContext.getTaskExecutions().createFromTaskId(task.id);
+    const newTaskExecution = await AppContext.getTaskExecutions().createFromTaskId(
+      StandardTracer.getSpanFromRequest(req),
+      task.id
+    );
     res.status(201).send(newTaskExecution.toJson());
   });
 }
