@@ -1,18 +1,14 @@
-import * as path from "path";
-import { Logger } from "../utils-std-ts/logger";
 import { FastifyInstance, RequestGenericInterface } from "fastify";
 import { UserPassword } from "../data/userPassword";
 import { AppContext } from "../appContext";
 import { Auth } from "../data/auth";
 import { User } from "../common-model/user";
-
-const logger = new Logger(path.basename(__filename));
+import { StandardTracer } from "../utils-std-ts/standardTracer";
 
 async function routes(fastify: FastifyInstance): Promise<void> {
   //
   fastify.get("/status/initialization", async (req, res) => {
-    logger.debug(`[${req.method}] ${req.url}`);
-    if ((await AppContext.getUsers().list()).length === 0) {
+    if ((await AppContext.getUsers().list(StandardTracer.getSpanFromRequest(req))).length === 0) {
       res.status(201).send({ initialized: false });
     } else {
       res.status(201).send({ initialized: true });
@@ -26,17 +22,16 @@ async function routes(fastify: FastifyInstance): Promise<void> {
     };
   }
   fastify.post<PostSession>("/session", async (req, res) => {
-    logger.info(`[${req.method}] ${req.url}`);
     if (!req.body.name) {
       return res.status(400).send({ error: "Missing: Name" });
     }
     if (!req.body.password) {
       return res.status(400).send({ error: "Missing: Password" });
     }
-    const user = await AppContext.getUsers().getByName(req.body.name);
+    const user = await AppContext.getUsers().getByName(StandardTracer.getSpanFromRequest(req), req.body.name);
     if (!user) {
       return res.status(403).send({ error: "Authentication Failed" });
-    } else if (await UserPassword.checkPassword(user, req.body.password)) {
+    } else if (await UserPassword.checkPassword(StandardTracer.getSpanFromRequest(req), user, req.body.password)) {
       res.status(201).send({ success: true, token: await Auth.generateJWT(user) });
     } else {
       return res.status(403).send({ error: "Authentication Failed" });
@@ -44,9 +39,8 @@ async function routes(fastify: FastifyInstance): Promise<void> {
   });
 
   fastify.get("/", async (req, res) => {
-    logger.debug(`[${req.method}] ${req.url}`);
     await Auth.mustBeAuthenticated(req, res);
-    const users = await AppContext.getUsers().list();
+    const users = await AppContext.getUsers().list(StandardTracer.getSpanFromRequest(req));
     for (const user of users) {
       delete user.passwordEncrypted;
     }
@@ -60,9 +54,8 @@ async function routes(fastify: FastifyInstance): Promise<void> {
     };
   }
   fastify.post<PostUser>("/", async (req, res) => {
-    logger.info(`[${req.method}] ${req.url}`);
     let isInitialized = true;
-    if ((await AppContext.getUsers().list()).length === 0) {
+    if ((await AppContext.getUsers().list(StandardTracer.getSpanFromRequest(req))).length === 0) {
       isInitialized = false;
     }
     const userSession = await Auth.getUserSession(req);
@@ -76,12 +69,12 @@ async function routes(fastify: FastifyInstance): Promise<void> {
     if (!req.body.password) {
       return res.status(400).send({ error: "Missing: Password" });
     }
-    if (await AppContext.getUsers().getByName(req.body.name)) {
+    if (await AppContext.getUsers().getByName(StandardTracer.getSpanFromRequest(req), req.body.name)) {
       return res.status(400).send({ error: "Username Already Exists" });
     }
     newUser.name = req.body.name;
-    await UserPassword.setPassword(newUser, req.body.password);
-    await AppContext.getUsers().add(newUser);
+    await UserPassword.setPassword(StandardTracer.getSpanFromRequest(req), newUser, req.body.password);
+    await AppContext.getUsers().add(StandardTracer.getSpanFromRequest(req), newUser);
     res.status(201).send({});
   });
 }
